+++ CHANGELOG.md
@@ -4,12 +4,12 @@
 
 ### Changed
 
-1. **BREAKING CHANGE** 
-   Reworked event struct in `get-status`, `get-event` and `get-events` response to include
+1. Reworked event struct in `get-status`, `get-event` and `get-events` response to include:
    - event type code and description
    - event reason code and description
+   - event direction code and description
 
-2. Added _en_ dictionary for event type and reason lookup.
+2. Added protocol version to configuration.
 
 ## [v0.8.0] - 2022-07-01